requirements.txt
binance-connector
python-dotenv
SQLAlchemy
pytest
